# internal/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API de disponibilite des salles campus (buildings, rooms, schedules, health).",
        "title": "Campus Room Status API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/rooms/{code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Get room detail and today's schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.RoomDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.ErrorResponse"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "api.EventResponse": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "organizer": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "api.RoomDetailResponse": {
            "type": "object",
            "properties": {
                "building": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "current_event": {
                    "$ref": "#/definitions/api.EventResponse"
                },
                "floor": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "next_event": {
                    "$ref": "#/definitions/api.EventResponse"
                },
                "schedule_today": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.EventResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}